refactor: 카테고리 랭킹 페이지 및 상품 그리드 구조 개선 및 dead code 정리
- ProductCard, ProductGrid, CategorySelector 등 중복/유사 컴포넌트 통합 및 공통화
- 랭킹/히스토리/일반 상품 그리드 UI/로직 일원화 (ProductGrid)
- 랭킹 뱃지 기능 ProductCard로 통합, RankingProductCard 등 dead code 제거
- CategorySelector를 공통 컴포넌트로 이동 및 재사용성 향상
- 사용하지 않는 import, 불필요한 파일/폴더, 중복 코드, dead code 일괄 삭제
- 각 컴포넌트 단일 책임 원칙 적용 및 구조 개선
- 전체적인 코드 일관성 및 유지보수성 향상
