+++ src/components/ranking/CategorySelector.tsx
@@ -33,7 +33,7 @@ export const CategorySelector = ({
           className={cn(
             'h-8 px-3 text-xs font-medium transition-colors',
             selectedCategory === category.value
-              ? 'hover:bg-primary-400 border-primary-300 bg-primary-300 text-white'
+              ? 'border-primary-300 bg-primary-300 text-text-on hover:opacity-80'
               : 'border-border-200 bg-bg-100 text-text-200 hover:bg-bg-200 hover:text-text-100',
           )}
         >
