+++ src/components/ranking/index.ts
@@ -1,3 +1,1 @@
-export { RankingProductCard } from './RankingProductCard';
 export { CategorySelector, type CategoryOption } from './CategorySelector';
-export { RankingProductGrid } from './RankingProductGrid';
